# implementation_examples/openapi.yaml
openapi: 3.1.0
...
info:
  title: Oolix API
  version: 0.2.0
  description: |
    Privacy-safe partner media activation control plane.

    **The rule every endpoint serves.** Oolix is the control plane; Data
    Partners remain the customer-data owners (§1). No request or response in
    this contract carries a `partner_user_id`, a segment membership list or an
    exact audience count. Reach is always a bucket (§72) and attribution is
    always an opaque token resolved server-side by SHA-256 (§90).
  contact:
    name: Oolix engineering
  license:
    name: Oolix proprietary
    identifier: LicenseRef-Oolix-Proprietary
servers:
  - url: http://localhost:4000
    description: Local development (docker compose stack)
  - url: '{scheme}://{host}'
    description: |
      Deployed environments. The host is supplied per environment rather than
      hard-coded, so a generated client can never default to someone else's
      domain.
    variables:
      scheme: { default: https, enum: [https] }
      host: { default: localhost:4000 }
...
security:
  - oidcBearer: []
paths:
...
  /v1/attribution/click/{token}:
    post:
      operationId: recordClick
      tags: [Attribution]
      summary: Record a click against an opaque token
      description: |
        §14, §90. The response is **identical** whether or not the token is
        recognised -- a differing status would let a caller probe for valid
        tokens.
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Attribution outcome
          content:
            application/json:
              schema:
                type: object
                properties:
                  attributed: { type: boolean }
                  activation_id: { type: string, format: uuid }

        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        §64: OIDC access token. Identity is delegated to the provider; Oolix
        stores only the subject claim. Roles are read from the database for the
        active organization, never from a token claim (§4.2).
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '§92.3: Oolix-issued Agent access token, 15-minute lifetime.'
    crmApiKey:
      type: http
      scheme: bearer
      description: "§71: the Buyer's CRM API key. Hashed at rest and rotatable."
...
  responses:
...
    RateLimited:
      description: |
        `SYS_001` -- §86's per-class budget for this caller is exhausted. The
        response carries `X-RateLimit-*`, `Retry-After` and
        `retry_after_seconds`, so a client can pace itself instead of guessing
        (§94).
      headers:
        X-RateLimit-Limit:
          { schema: { type: integer }, description: Requests allowed in the window. }
        X-RateLimit-Remaining:
          { schema: { type: integer }, description: Requests left in the window. }
        X-RateLimit-Reset:
          { schema: { type: integer }, description: Unix seconds when the window frees up. }
        Retry-After: { schema: { type: integer }, description: Seconds to wait before retrying. }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      description: |
        §53, §99. Every error uses this envelope: a canonical code, a human
        message, an explicit retryable flag, field errors and the correlation
        id. §99 forbids exposing stack traces or secrets.
      properties:
        error:
          type: object
          required: [code, message, retryable, correlation_id, field_errors]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            retryable: { type: boolean }
            correlation_id: { type: string }
            retry_after_seconds: { type: integer }
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    ErrorCode:
      type: string
      description: '§77.2 canonical error registry.'
      enum:
        [
          AUTH_001,
          PERM_001,
          PERM_002,
          VAL_001,
          CAMP_001,
          CAMP_002,
          CAMP_003,
          PART_001,
          PART_002,
          CHAN_001,
          CHAN_002,
          AGENT_001,
          SYS_001,
          SYS_002,
          IDEMPOTENCY_CONFLICT,
        ]

    # --- §89 canonical channel enum ------------------------------------------
...